bootstrap metric — increase size from uint16 to uint32

+++ db-schema/ij/report.sql
@@ -22,7 +22,7 @@ create table report2
   `build_c2`                   UInt16 CODEC(DoubleDelta, ZSTD(20)),
   `build_c3`                   UInt16 CODEC(DoubleDelta, ZSTD(20)),
 
-  `bootstrap_d`          UInt16 CODEC (Gorilla, ZSTD(20)),
+  `bootstrap_d`          UInt32 CODEC (Gorilla, ZSTD(20)),
   `appInitPreparation_d` UInt16 CODEC (Gorilla, ZSTD(20)),
   `appInit_d`            UInt16 CODEC (Gorilla, ZSTD(20)),
